+++ src/main/java/com/tlabs/eve/api/character/CharacterAssetsRequest.java
@@ -6,7 +6,7 @@ import com.tlabs.eve.api.AssetListResponse;
 
 public final class CharacterAssetsRequest extends CharacterRequest<AssetListResponse> {
 
-    public static final int MASK = 2;
+    public static final long MASK = 2;
 
     public CharacterAssetsRequest(String characterID) {
         super(AssetListResponse.class, "/char/AssetList.xml.aspx", MASK, characterID);
