+++ src/main/java/com/tlabs/eve/api/character/CharacterContractsRequest.java
@@ -5,7 +5,7 @@ package com.tlabs.eve.api.character;
 import com.tlabs.eve.api.ContractListResponse;
 
 public class CharacterContractsRequest extends CharacterRequest<ContractListResponse> {
-    public static final int MASK = 67108864;
+    public static final long MASK = 67108864;
 
     public CharacterContractsRequest(final String charID) {
         super(ContractListResponse.class, "/char/Contracts.xml.aspx", MASK, charID);
