+++ src/main/java/com/tlabs/eve/api/character/CharacterItemLocationRequest.java
@@ -7,7 +7,7 @@ import com.tlabs.eve.api.ItemLocationResponse;
 
 public final class CharacterItemLocationRequest extends CharacterRequest<ItemLocationResponse> {
 
-    public static final int MASK = EveAPI.CHAR_FULL; //FIXME CAK not known yet
+    public static final long MASK = EveAPI.CHAR_FULL; //FIXME CAK not known yet
 
     public CharacterItemLocationRequest(String charID, long[] ids) {
         super(ItemLocationResponse.class, "/char/Locations.xml.aspx", MASK, charID);
