+++ src/main/java/com/tlabs/eve/api/character/CharacterKillLogRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.character;
 import com.tlabs.eve.api.mail.KillLogResponse;
 
 public final class CharacterKillLogRequest extends CharacterRequest<KillLogResponse> {
-    public static final int MASK = 256;
+    public static final long MASK = 256;
 
     public CharacterKillLogRequest(final String charID) {
         super(KillLogResponse.class, "/char/KillLog.xml.aspx", MASK, charID);
