+++ src/main/java/com/tlabs/eve/api/character/CharacterMarketOrderRequest.java
@@ -5,7 +5,7 @@ package com.tlabs.eve.api.character;
 import com.tlabs.eve.api.MarketOrderResponse;
 
 public final class CharacterMarketOrderRequest extends CharacterRequest<MarketOrderResponse> {
-    public static final int MASK = 4096;
+    public static final long MASK = 4096;
 
     public CharacterMarketOrderRequest(String characterID) {
         super(MarketOrderResponse.class, "/char/MarketOrders.xml.aspx", MASK, characterID);
