+++ src/main/java/com/tlabs/eve/api/character/CharacterSheetRequest.java
@@ -6,7 +6,7 @@ package com.tlabs.eve.api.character;
 
 
 public final class CharacterSheetRequest extends CharacterRequest<CharacterSheetResponse> {
-    public static final int MASK = 8;
+    public static final long MASK = 8;
 
     public CharacterSheetRequest(String charID) {
         super(CharacterSheetResponse.class, "/char/CharacterSheet.xml.aspx", MASK, charID);
