+++ src/main/java/com/tlabs/eve/api/corporation/CorporationAccountBalanceRequest.java
@@ -5,7 +5,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.AccountBalanceResponse;
 
 public final class CorporationAccountBalanceRequest extends CorporationRequest<AccountBalanceResponse> {
-    public static final int MASK = 1;
+    public static final long MASK = 1;
 
     public CorporationAccountBalanceRequest(String corpID) {
         super(AccountBalanceResponse.class, "/corp/AccountBalance.xml.aspx", MASK, corpID);
