+++ src/main/java/com/tlabs/eve/api/corporation/CorporationContactListRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.mail.ContactListResponse;
 
 public final class CorporationContactListRequest extends CorporationRequest<ContactListResponse> {
-    public static final int MASK = 16;
+    public static final long MASK = 16;
 
     public CorporationContactListRequest(final String corporationID) {
         super(ContactListResponse.class, "/corp/ContactList.xml.aspx", MASK, corporationID);
