+++ src/main/java/com/tlabs/eve/api/corporation/CorporationContractBidsRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.ContractBidsResponse;
 
 public final class CorporationContractBidsRequest extends CorporationRequest<ContractBidsResponse> {
-    public static final int MASK = 8388608;
+    public static final long MASK = 8388608;
 
     public CorporationContractBidsRequest(final String corpID) {
         super(ContractBidsResponse.class, "/corp/ContractBids.xml.aspx", MASK, corpID);
