+++ src/main/java/com/tlabs/eve/api/corporation/CorporationKillLogRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.mail.KillLogResponse;
 
 public final class CorporationKillLogRequest extends CorporationRequest<KillLogResponse> {
-    public static final int MASK = 256;
+    public static final long MASK = 256;
 
     public CorporationKillLogRequest(final String corporationID) {
         super(KillLogResponse.class, "/corp/KillLog.xml.aspx", MASK, corporationID);
