+++ src/main/java/com/tlabs/eve/api/corporation/CorporationMarketOrderRequest.java
@@ -5,7 +5,7 @@ package com.tlabs.eve.api.corporation;
 import com.tlabs.eve.api.MarketOrderResponse;
 
 public final class CorporationMarketOrderRequest extends CorporationRequest<MarketOrderResponse> {
-    public static final int MASK = 4096;
+    public static final long MASK = 4096;
 
     public CorporationMarketOrderRequest(String corpID) {
         super(MarketOrderResponse.class, "/corp/MarketOrders.xml.aspx", MASK, corpID);
