+++ src/main/java/com/tlabs/eve/api/corporation/CorporationStandingsRequest.java
@@ -5,7 +5,7 @@ import com.tlabs.eve.api.StandingsResponse;
 
 
 public class CorporationStandingsRequest extends CorporationRequest<StandingsResponse> {
-    public static final int MASK = 262144;
+    public static final long MASK = 262144;
 
     public CorporationStandingsRequest(String corpID) {
         super(StandingsResponse.class, "/corp/Standings.xml.aspx", MASK, corpID);
