+++ src/main/java/com/tlabs/eve/api/corporation/MemberTrackingRequest.java
@@ -3,7 +3,7 @@ package com.tlabs.eve.api.corporation;
 
 
 public final class MemberTrackingRequest extends CorporationRequest<MemberTrackingResponse> {
-    public static final int MASK = 2048;
+    public static final long MASK = 2048;
 
     public MemberTrackingRequest(String corporationID, boolean extended) {
         super(MemberTrackingResponse.class, "/corp/MemberTracking.xml.aspx", MASK, corporationID);
