+++ src/main/java/com/tlabs/eve/api/mail/MailMessagesRequest.java
@@ -6,7 +6,7 @@ import com.tlabs.eve.api.character.CharacterRequest;
 
 ///char/MailMessages.xml.aspx
 public final class MailMessagesRequest extends CharacterRequest<MailMessagesResponse> {
-    public static final int MASK = 2048;
+    public static final long MASK = 2048;
 
     public MailMessagesRequest(String characterID) {
         super(MailMessagesResponse.class, "/char/MailMessages.xml.aspx", MASK, characterID);
